Update playerParse and playerController
This update will Create a new Player object from the Player prefab and name it the new players name. It will also Get the playerController and assign anything new to the player

This is done in the newPlayer function of controllerParse so when a new player joins a new Player object is made and will be callable from its own unique script that will handle all the players information.