Fix Call and Raise functions
There was an issue with the two functions this commit will fix them with new logic so that "currentBet" is the current bet that players either have to match, raise, or fold.

Note: pressing the bet button will automatically raise you by the amount given if it isnt the initial bet.