Fix nextTurn, Call,  and Fold functions
Call - There was a bug that didnt check if the highestBidder was calling as they get the turn to raise the bid if they so choose.

Fold - Fixed a bug where folding would skip the next players turn. Also added an escape for when newRound() ends.

nextTurn() - Fixed issue where folding would skip the next players turn.

newRound() - reset some of the players stats so that the next round of betting starts from 0.