Implement blinds
This will add in functionality to play with a big and small blind. Betting after the flop will resume with the original order rather than with after the forced bets.

Also fixed a bug with the call function where it would call the current bet which after the reveals would be more than 0 which players may not wish to bet any amount.